A newly created topic will not be returned at the top of the list, even it the post wasn't edited yet.
This is due to a bug, that the initial root post was created with the fields 'created_at' and last_touch' as null. Thus they were sorted to the end of the topic list upon retrieval. I refactored the creation code to be using the same function as the post_edit rpc function.